Up to Chapter 18.3, re-migrated everything, add method not working correctly, but seems like it's because of the add category

+++ CodingEvents/ViewModels/AddEventViewModel.cs
@@ -20,7 +20,7 @@ namespace CodingEvents.ViewModels
         public string Location { get; set; }
 
         [Range(0, 100000, ErrorMessage ="Number of attendees must be between 0 and 100,000")]
-        public double Attendees { get; set; }
+        public int Attendees { get; set; }
 
         [Required(ErrorMessage ="Category is required.")]
         public int CategoryId { get; set; }
